Moved bootstrapOutboundChannel to the responseGenerator-class
