+++ proxy-config/src/main/resources/reference.conf
@@ -9,6 +9,16 @@ host.localhost = {
         idp: google
       },
       {
+        path: /"?"
+        security: 3
+        idp: idporten
+      },
+      {
+        path: /fakulteter
+        security: 3
+        idp: idporten
+      },
+      {
         path: /idporten
         security: 3
         idp: idporten
