+++ proxy-proxy/src/main/java/no/difi/idporten/oidc/proxy/proxy/ResponseGenerator.java
@@ -1,18 +1,21 @@
 package no.difi.idporten.oidc.proxy.proxy;
 
 import com.google.gson.Gson;
+import io.netty.bootstrap.Bootstrap;
+import io.netty.buffer.PooledByteBufAllocator;
 import io.netty.buffer.Unpooled;
-import io.netty.channel.ChannelFutureListener;
-import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.*;
 import io.netty.handler.codec.http.*;
 import io.netty.util.AsciiString;
 import io.netty.util.CharsetUtil;
+import no.difi.idporten.oidc.proxy.api.CookieStorage;
 import no.difi.idporten.oidc.proxy.api.IdentityProvider;
 import no.difi.idporten.oidc.proxy.api.ProxyCookie;
 import no.difi.idporten.oidc.proxy.lang.IdentityProviderException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.net.SocketAddress;
 import java.util.HashMap;
 
 public class ResponseGenerator {
@@ -51,8 +54,7 @@ public class ResponseGenerator {
      * Default response for when nothing is configured for the host
      */
     protected void generateDefaultResponse(ChannelHandlerContext ctx, String host) {
-        StringBuilder content = new StringBuilder();
-        content.append(String.format("no cannot use %s", host));
+        String content = String.format("no cannot use %s", host);
         FullHttpResponse response = new DefaultFullHttpResponse(
                 HttpVersion.HTTP_1_1,
                 HttpResponseStatus.BAD_REQUEST,
@@ -82,8 +84,81 @@ public class ResponseGenerator {
                 "%s; %s=%s", APPLICATION_JSON, HttpHeaderValues.CHARSET, CharsetUtil.UTF_8));
         logger.debug("Setting Set-Cookie to the response");
         CookieHandler.insertCookieToResponse(result, proxyCookieObject.getName(), proxyCookieObject.getUuid());
+        System.out.println();
         logger.debug(String.format("Created JWT response:\n%s", result));
         ctx.writeAndFlush(result).addListener(ChannelFutureListener.CLOSE);
+    }
+
+    /**
+     * This is what happens when the proxy needs to work as a normal proxy.
+     * We could also direct IDP traffic this way instead of the the apache.http.HttpClient, but then we would need
+     * SSL set up.
+     *
+     * @param ctx
+     * @param outboundAddress
+     * @param httpRequest
+     */
+
+    public Channel generateProxyResponse(ChannelHandlerContext ctx, SocketAddress outboundAddress, HttpRequest httpRequest){
+        Channel outboundChannel;
+        logger.info(String.format("Bootstrapping channel %s", ctx.channel()));
+        final Channel inboundChannel = ctx.channel();
+
+
+        Bootstrap b = new Bootstrap();
+        b.group(inboundChannel.eventLoop()).channel(ctx.channel().getClass());
+        b.handler(new OutboundInitializer(inboundChannel))
+                .option(ChannelOption.AUTO_READ, false);
+
+        b.option(ChannelOption.ALLOCATOR, PooledByteBufAllocator.DEFAULT);
+        b.option(ChannelOption.TCP_NODELAY, true);
+        b.option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 15000);
+
+        b.option(ChannelOption.SO_SNDBUF, 1048576);
+        b.option(ChannelOption.SO_RCVBUF, 1048576);
+
+        ChannelFuture f = b.connect(outboundAddress);
+
+        outboundChannel = f.channel();
+        logger.debug(String.format("Made outbound channel: %s", outboundChannel));
+        f.addListener(new ChannelFutureListener() {
+            @Override
+            public void operationComplete(ChannelFuture future) throws Exception {
+                logger.debug("Outbound channel operation complete");
+                if (future.isSuccess()) {
+                    // connection complete start to read first data
+                    logger.debug("Outbound channel operation success");
+                    outboundChannel.writeAndFlush(httpRequest).addListener(new ChannelFutureListener() {
+                        @Override
+                        public void operationComplete(ChannelFuture future) throws Exception {
+                            if (future.isSuccess()) {
+                                // was able to flush out data, start to read the next chunk
+                                ctx.channel().read();
+                            } else {
+                                future.channel().close();
+                            }
+                        }
+                    });
+                } else {
+                    // Close the connection if the connection attempt has failed.
+                    logger.debug("Outbound channel operation failure");
+                    inboundChannel.close();
+                }
+            }
+        });
+        return outboundChannel;
+
+    }
+
+    protected Channel generateSecuredResponse(ChannelHandlerContext ctx, SocketAddress outboundAddress,
+                                              HttpRequest httpRequest, HashMap<String, String> userData,
+                                              ProxyCookie proxyCookieObject) throws  IdentityProviderException{
+
+        System.out.println("HEADER:"+httpRequest.headers().entries());
+
+
+
+        return generateProxyResponse(ctx, outboundAddress, httpRequest);
     }
 
 }
